# docex/test_projects/elastic/infra/contracts/api.web.rest.openapi.yml
# The `rest` surface of the `api.web` core service.
#
# The path is <codebase>.<service>.<surface>.<format>.<ext> — keyed on the
# SURFACE, not just the core service: one core service may declare several
# boundaries with unrelated consumers and each gets its own document. This file
# is `api.web`'s single surface, `surfaces: {rest: {api_styles: [rest]}}` in
# infra.yml, and the format follows from that style rather than from the core
# service's `role` (cicl.md § Surfaces).
#
# docex check's contract gate verifies that paths declared here exist in the
# running core service (provider-side contract testing), and its orphan arm
# fails on any file in this folder that parses as a contract but matches no
# declared surface — so nothing may be left behind here.
#
# WHY `GET /health` BELONGS HERE, and only here. It is not because every
# long-running core service owes a /health route — none of them do; a non-`web`
# core service's liveness is its container probe, `./health.sh <service>`.
# `api.web` is on the `web` network AND declares an `openapi` surface, which is
# the one case healthchecks.md § web services also serve GET /health describes,
# and it is what docex check's contract_health_path gate asserts. The route
# exists for the reverse proxy (`health_check_path: /health`), not for a
# reader of this contract.
#
# WHY THE /diagnostics/* PATHS ARE NOT A HEALTH FAN-OUT. They are `api.web`
# probing the reachability of two BACKING services — `probe` (sidecar/nginx)
# and `events` (analytics_db/clickhouse) — and they report on nothing that is a
# core service. There is no fan-out anywhere in this project and none is
# possible: healthchecks.md § What this doctrine does not do forbids one
# service reporting on another's liveness. These two routes used to sit under
# /health/* beside a fan-out endpoint that has been deleted, and they were
# moved out precisely so that no future reader mistakes them for its remnant.

openapi: "3.2.0"
info:
  title: docex_smoke_elastic.api.web
  version: "0.0.1"
  description: |
    Test-project web core service for the doctrine smoke test. POST a
    payload to /pings, the `api.worker` core service picks it up via
    postgres. /diagnostics/{probe,events} exercise the project-local sidecar
    + analytics_db backings. POST /jobs/drain asks `api.worker` to drain the
    deferred-job queue now, across that core service's `rpc` surface.
servers:
  - url: "https://docex-smoke-elastic.luxrnd.tech"
    description: prod (bare-project → prod's bare env → domain_default_service)
  - url: "https://api-web.prod.docex-smoke-elastic.luxrnd.tech"
    description: prod (canonical per-service host)
  - url: "https://stage.docex-smoke-elastic.luxrnd.tech"
    description: stage (bare-env → domain_default_service)
  - url: "https://api-web.stage.docex-smoke-elastic.luxrnd.tech"
    description: stage (canonical per-service host)

paths:
  /health:
    get:
      summary: Service liveness.
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [version]
                properties:
                  version:
                    type: string
                    example: "0.0.1"

  /diagnostics/probe:
    get:
      summary: Backing-service reachability — the nginx sidecar over the internal network.
      operationId: diagnosticsProbe
      responses:
        "200":
          description: probe reachable
          content:
            application/json:
              schema:
                type: object
                required: [reachable]
                properties:
                  reachable:
                    type: string
                    example: "true"
                  status_code:
                    type: integer
        "503":
          description: probe unreachable

  /diagnostics/events:
    get:
      summary: Backing-service reachability — ClickHouse (analytics_db) on its native port.
      operationId: diagnosticsEvents
      responses:
        "200":
          description: events backing reachable
          content:
            application/json:
              schema:
                type: object
                required: [reachable]
                properties:
                  reachable:
                    type: string
                    example: "true"
                  port:
                    type: integer
        "503":
          description: events backing unreachable

  /pings:
    post:
      summary: Enqueue a ping for the worker to process.
      operationId: createPing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payload]
              properties:
                payload:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
                    format: uuid
        "400":
          description: Invalid payload

  # The two job-deferral routes. They call the SAME `ContJobs` driving port
  # that `api.clock` fires on a cron tick, which is the side effect
  # clock.md § Architecture calls out: firing a scheduled job by hand in
  # `dev` is an ordinary HTTP call rather than a special path.
  #
  # ONE ROUTE PER JOB, deliberately. `POST /jobs/{name}` would be shorter
  # and would make this contract meaningless — "some job, named at
  # runtime" describes nothing a consumer can rely on and nothing docex
  # check's contract gate can verify.
  #
  # 202, not 201: nothing has happened yet. The response is a receipt for
  # a queued row, and `api.worker` performs the work on a later pass.
  /jobs/prune_pings:
    post:
      summary: Defer a prune of expired processed pings.
      description: |
        Enqueues the `prune_pings` job onto the `jobs` table and returns
        immediately. `api.worker` claims it on its next pass and performs
        the retention delete. Also fired by `api.clock` at 03:00 UTC daily.
      operationId: deferPrunePings
      responses:
        "202":
          description: Accepted — the job was enqueued, not performed
          content:
            application/json:
              schema:
                type: object
                required: [job_id]
                properties:
                  job_id:
                    type: string
                    format: uuid
        "503":
          description: the queue could not be written; nothing was deferred

  /jobs/heartbeat:
    post:
      summary: Defer a no-op liveness job.
      description: |
        Enqueues the `heartbeat` job onto the `jobs` table and returns
        immediately. The job performs nothing; it exists so the
        fire -> defer -> drain path is observable. Also fired by
        `api.clock` every minute.
      operationId: deferHeartbeat
      responses:
        "202":
          description: Accepted — the job was enqueued, not performed
          content:
            application/json:
              schema:
                type: object
                required: [job_id]
                properties:
                  job_id:
                    type: string
                    format: uuid
        "503":
          description: the queue could not be written; nothing was deferred

  # The one route on this surface that crosses to another core service. It is
  # 200 rather than 202 because, unlike the two deferral routes above, the work
  # HAS happened by the time this replies — the worker's own reply travels back
  # out through this edge carrying the count.
  /jobs/drain:
    post:
      summary: Ask `api.worker` to drain the deferred-job queue now.
      description: |
        Calls `api.worker`'s `rpc` surface (`POST /drain`, described by
        api.worker.rpc.asyncapi.yml) and returns the number of jobs that core
        service performed.

        WHY this crosses a process boundary instead of draining here: the
        perform side of the queue belongs to `api.worker`. An edge that
        drained the queue in its own process, with its own sizing and its own
        request-scoped lifetime, would be performing rather than deferring —
        the same violation clock.md § The clock defers; it does not work
        forbids of `api.clock`. This edge asks; the worker performs.

        `performed: 0` IS A SUCCESS. The worker's own poll loop drains on its
        interval, so finding the queue already empty is the ordinary outcome
        of asking at the wrong moment. A caller that treats 0 as a failure is
        asserting on scheduling, which this route does not promise.
      operationId: drainJobs
      responses:
        "200":
          description: the worker drained the queue and reported the count
          content:
            application/json:
              schema:
                type: object
                required: [performed]
                properties:
                  performed:
                    type: integer
                    minimum: 0
                    example: 3
        "503":
          description: the worker could not be reached; nothing was drained